Grouping shape command into different files
